+++ js/search_filters.js
@@ -5,6 +5,8 @@ function Search_Filters_GUI() {
     // filter objects set
     THAT.filters = new Object();
 
+    THAT.$issue_load = undefined;
+
     THAT.create_filters = function(filters_array)
     {
         for ( var i = 0 ; i < filters_array.length ; i++ )
@@ -25,7 +27,19 @@ function Search_Filters_GUI() {
                 console.warn('Filter not yet implemented');
                 break;
         }
+    }
 
+    THAT.init_category_events = function() {
+        THAT.filters.filter_by_category.gui_widget.jq_pointer.unbind().on('change', function(){
+            if ( $(this).hasClass("disabled") ) return;
+            JOM.search_filters.update_issues();
+        });
+    }
+
+    THAT.update_issues = function() {
+        THAT.$issue_load.fadeIn();
+        THAT.filters.filter_by_issue.parent_id = THAT.filters.filter_by_category.gui_widget.jq_pointer.val();
+        THAT.filters.filter_by_issue.load();
     }
 
     THAT.init_filter = function(filter_name, nonce_obj, $jQ_ptr)
@@ -34,5 +48,10 @@ function Search_Filters_GUI() {
         THAT.filters[filter_name].jq_pointer    = $jQ_ptr;
         THAT.filters[filter_name].load();
     }
+
+    // constructor
+        THAT.$issue_load    = $("#jom_filter_by_issue").next();
+        THAT.$issue_load.hide();
+    // end constructor
 
 }